grammar, spelling, and layout
I fixed some grammar and spelling. I also made a change to the layout in one section on line 215, the run on of the sentence seemed to fit better in a list. Some of the paragraphs are hard to change because I may not know what you were going for when you wrote the sentence or paragraph, so my changes are in the way I interpreted what you are trying to convey.